src: video,stream: Move CaptureConfiguration struct
